Move msbuild pattern test helper
