+++ buildpacks/dotnet/tests/dotnet_publish_test.rs
@@ -1,7 +1,6 @@
-use crate::tests::{default_build_config, get_rid};
+use crate::tests::{default_build_config, get_rid, replace_msbuild_log_patterns_with_placeholder};
 use indoc::{formatdoc, indoc};
 use libcnb_test::{assert_contains, assert_empty, PackResult, TestRunner};
-use regex::Regex;
 
 #[test]
 #[ignore = "integration test"]
@@ -117,31 +116,4 @@ fn test_dotnet_publish_with_global_json_and_custom_verbosity_level() {
             );
         },
     );
-}
-
-fn replace_msbuild_log_patterns_with_placeholder(input: &str, placeholder: &str) -> String {
-    // Define regex patterns for dynamic/undeterministic msbuild log output to replace for simple integration test assertions.
-    let patterns = vec![
-        // Date-time pattern
-        r"\d{2}/\d{2}/\d{4} \d{2}:\d{2}:\d{2}",
-        // Elapsed time pattern
-        r"\d{2}:\d{2}:\d{2}\.\d{2}",
-        // Server message with UUID pattern
-        r"server - server processed compilation - [0-9a-fA-F-]{36}",
-        // Parentheses text pattern (contains elapsed time in various forms)
-        r"\(in [^)]+\)",
-        // Milliseconds pattern
-        r"\b\d+ms\b",
-        // Section between _CopyOutOfDateSourceItemsToOutputDirectory and _CopyResolvedFilesToPublishAlways pattern:
-        // (Log entries in these sections are not written deterministically).
-        r"(?s)_CopyOutOfDateSourceItemsToOutputDirectory:.*?_CopyResolvedFilesToPublishAlways:",
-    ];
-
-    let mut result = input.to_string();
-    for pattern in patterns {
-        let regex = Regex::new(pattern).unwrap();
-        result = regex.replace_all(&result, placeholder).to_string();
-    }
-
-    result
 }
