+++ buildpacks/dotnet/tests/mod.rs
@@ -1,4 +1,5 @@
 use libcnb_test::BuildConfig;
+use regex::Regex;
 use std::path::Path;
 
 mod detect_test;
@@ -28,4 +29,31 @@ fn get_dotnet_arch() -> String {
     let arch = "arm64";
 
     arch.to_string()
+}
+
+fn replace_msbuild_log_patterns_with_placeholder(input: &str, placeholder: &str) -> String {
+    // Define regex patterns for dynamic/undeterministic msbuild log output to replace for simple integration test assertions.
+    let patterns = vec![
+        // Date-time pattern
+        r"\d{2}/\d{2}/\d{4} \d{2}:\d{2}:\d{2}",
+        // Elapsed time pattern
+        r"\d{2}:\d{2}:\d{2}\.\d{2}",
+        // Server message with UUID pattern
+        r"server - server processed compilation - [0-9a-fA-F-]{36}",
+        // Parentheses text pattern (contains elapsed time in various forms)
+        r"\(in [^)]+\)",
+        // Milliseconds pattern
+        r"\b\d+ms\b",
+        // Section between _CopyOutOfDateSourceItemsToOutputDirectory and _CopyResolvedFilesToPublishAlways pattern:
+        // (Log entries in these sections are not written deterministically).
+        r"(?s)_CopyOutOfDateSourceItemsToOutputDirectory:.*?_CopyResolvedFilesToPublishAlways:",
+    ];
+
+    let mut result = input.to_string();
+    for pattern in patterns {
+        let regex = Regex::new(pattern).unwrap();
+        result = regex.replace_all(&result, placeholder).to_string();
+    }
+
+    result
 }
